🚚 Move ExperienceStepFormState to data model package instead of UI
